corrected Path name, corrected ID generation method, corrected erb file names, separated memo class from app.rb
