+++ views/details.erb
@@ -6,9 +6,9 @@
   <%= @contents[:body] %>
 </p>
 <p>
-  <button type="button"onclick="location.href='/edit/<%= @contents[:id] %>'">変更 </button>
+  <button type="button"onclick="location.href='/memos/<%= @contents[:id] %>/edit'">変更 </button>
 </p>
-<form method="post" action="/show/<%= @contents[:id] %>">
+<form method="post" action="/memos/<%= @contents[:id] %>">
   <p>
     <input type="hidden" name="_method" value="DELETE">
     <input type="submit" value="削除">
