+++ views/new.erb
@@ -1,6 +1,6 @@
 <h1>メモの新規作成</h1>
 
-<form action="/add" method="post">
+<form action="/memos/new" method="post">
   <p>メモのタイトル<br>
   <input type="text" name="title"></p>
   <p>メモの詳細<br>
